+++ python/ray/actor.py
@@ -5,7 +5,7 @@ import weakref
 import ray.ray_constants as ray_constants
 import ray._raylet
 import ray._private.signature as signature
-from ray._private.runtime_env.validation import ParsedRuntimeEnv
+from ray.runtime_env import RuntimeEnv
 import ray.worker
 from ray.util.annotations import PublicAPI
 from ray.util.placement_group import configure_placement_group_based_on_context
@@ -450,7 +450,7 @@ class ActorClass:
             if isinstance(runtime_env, str):
                 new_runtime_env = runtime_env
             else:
-                new_runtime_env = ParsedRuntimeEnv(runtime_env).serialize()
+                new_runtime_env = RuntimeEnv(**runtime_env).serialize()
         else:
             new_runtime_env = None
 
@@ -498,7 +498,7 @@ class ActorClass:
             if isinstance(runtime_env, str):
                 new_runtime_env = runtime_env
             else:
-                new_runtime_env = ParsedRuntimeEnv(runtime_env).serialize()
+                new_runtime_env = RuntimeEnv(**runtime_env).serialize()
         else:
             new_runtime_env = None
 
@@ -586,7 +586,7 @@ class ActorClass:
             if isinstance(runtime_env, str):
                 new_runtime_env = runtime_env
             else:
-                new_runtime_env = ParsedRuntimeEnv(runtime_env or {}).serialize()
+                new_runtime_env = RuntimeEnv(**(runtime_env or {})).serialize()
         else:
             # Keep the new_runtime_env as None.  In .remote(), we need to know
             # if runtime_env is None to know whether or not to fall back to the
@@ -948,7 +948,7 @@ class ActorClass:
             if isinstance(runtime_env, str):
                 # Serialzed protobuf runtime env from Ray client.
                 new_runtime_env = runtime_env
-            elif isinstance(runtime_env, ParsedRuntimeEnv):
+            elif isinstance(runtime_env, RuntimeEnv):
                 new_runtime_env = runtime_env.serialize()
             else:
                 raise TypeError(f"Error runtime env type {type(runtime_env)}")
