+++ python/ray/runtime_context.py
@@ -1,7 +1,7 @@
 import ray.worker
 import logging
 from ray._private.client_mode_hook import client_mode_hook
-from ray._private.runtime_env.validation import ParsedRuntimeEnv
+from ray.runtime_env import RuntimeEnv
 from ray.util.annotations import PublicAPI
 
 logger = logging.getLogger(__name__)
@@ -164,13 +164,14 @@ class RuntimeContext(object):
 
     @property
     def runtime_env(self):
-        """Get the runtime env dict used for the current driver or worker.
+        """Get the runtime env used for the current driver or worker.
 
         Returns:
-            The runtime env dict currently using by this worker.
+            The runtime env currently using by this worker. The type of
+                return value is ray.runtime_env.RuntimeEnv.
         """
 
-        return ParsedRuntimeEnv.deserialize(self.get_runtime_env_string())
+        return RuntimeEnv.deserialize(self.get_runtime_env_string())
 
     @property
     def current_actor(self):
